Adding more validation, minor edits and tidying up, removing debug logs

+++ Spring-Boot-API/src/main/java/com/projekt/Spring_Boot_API/services/FolderService.java
@@ -1,41 +1,38 @@
 package com.projekt.Spring_Boot_API.services;
 
+import com.projekt.Spring_Boot_API.dtos.folder.FolderContentsDTO;
 import com.projekt.Spring_Boot_API.exceptions.folder.FolderNameEmptyException;
 import com.projekt.Spring_Boot_API.exceptions.folder.FolderNotFoundException;
+import com.projekt.Spring_Boot_API.exceptions.folder.OwnerFolderMismatchException;
 import com.projekt.Spring_Boot_API.exceptions.folder.UnauthorizedFolderActionException;
-import com.projekt.Spring_Boot_API.exceptions.user.UserNotFoundException;
 import com.projekt.Spring_Boot_API.models.Folder;
 import com.projekt.Spring_Boot_API.models.User;
 import com.projekt.Spring_Boot_API.repositories.IFolderRepository;
-import com.projekt.Spring_Boot_API.repositories.IUserRepository;
+import com.projekt.Spring_Boot_API.repositories.IItemRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class FolderService {
-    private final IUserRepository userRepository;
     private final IFolderRepository folderRepository;
+    private final IItemRepository itemRepository;
 
-    public Folder createFolder(String folderName, UUID parentFolderId, UUID ownerId) {
-        Folder parentFolder = null;
+    public Folder createFolder(String folderName, UUID parentFolderId) {
+        Folder parentFolder = folderRepository.findByFolderId(parentFolderId)
+                .orElseThrow(FolderNotFoundException::new);
 
-        if (parentFolderId != null) {
-            parentFolder = folderRepository.findByFolderId(parentFolderId)
-                    .orElseThrow(() -> new RuntimeException("Parent folder not found"));
-        }
-
-        User owner = userRepository.findByUserId(ownerId)
-                .orElseThrow(UserNotFoundException::new);
+        User user = authenticateUser();
+        checkFolderOwnership(user, parentFolder);
 
         if (folderName == null || folderName.isBlank()) {
             throw new FolderNameEmptyException();
         }
 
-        Folder folder = new Folder(folderName, parentFolder, owner);
+        Folder folder = new Folder(folderName, parentFolder, user);
 
         return folderRepository.save(folder);
     }
@@ -45,12 +42,17 @@ public class FolderService {
                 .orElseThrow(FolderNotFoundException::new);
 
         if (folder.getParentFolder() == null) {
-            throw new UnauthorizedFolderActionException("Root folder cannot be modified.");
+            throw new UnauthorizedFolderActionException();
         }
+
+        User user = authenticateUser();
+
+        checkFolderOwnership(user, folder);
 
         if (parentFolderId != null) {
             Folder parentFolder = folderRepository.findByFolderId(parentFolderId)
                     .orElseThrow(FolderNotFoundException::new);
+            checkFolderOwnership(user, parentFolder);
             folder.setParentFolder(parentFolder);
         }
 
@@ -66,16 +68,37 @@ public class FolderService {
                 .orElseThrow(FolderNotFoundException::new);
 
         if (folder.getParentFolder() == null) {
-            throw new UnauthorizedFolderActionException("Root folder cannot be deleted.");
+            throw new UnauthorizedFolderActionException();
         }
+
+        checkFolderOwnership(authenticateUser(), folder);
 
         folderRepository.delete(folder);
     }
 
-    public List<Folder> getSubFolders(UUID parentFolderId) {
+    public FolderContentsDTO getContents(UUID parentFolderId) {
         Folder parentFolder = folderRepository.findByFolderId(parentFolderId)
                 .orElseThrow(FolderNotFoundException::new);
 
-        return folderRepository.findByParentFolder(parentFolder);
+        checkFolderOwnership(authenticateUser(), parentFolder);
+
+        return new FolderContentsDTO(
+                folderRepository.findByParentFolder(parentFolder),
+                itemRepository.findByFolder(parentFolder)
+        );
+    }
+
+    private User authenticateUser() {
+        return (User) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+    }
+
+    private void checkFolderOwnership(User user, Folder folder) {
+        if (!user.getUserId()
+                .equals(folder.getUser().getUserId())) {
+            throw new OwnerFolderMismatchException();
+        }
     }
 }
