+++ Spring-Boot-API/src/main/java/com/projekt/Spring_Boot_API/services/ItemService.java
@@ -1,7 +1,9 @@
 package com.projekt.Spring_Boot_API.services;
 
 import com.projekt.Spring_Boot_API.exceptions.folder.FolderNameEmptyException;
+import com.projekt.Spring_Boot_API.exceptions.folder.OwnerFolderMismatchException;
 import com.projekt.Spring_Boot_API.exceptions.item.ItemNotFoundException;
+import com.projekt.Spring_Boot_API.exceptions.item.OwnerItemMismatchException;
 import com.projekt.Spring_Boot_API.exceptions.user.UserNotFoundException;
 import com.projekt.Spring_Boot_API.models.Folder;
 import com.projekt.Spring_Boot_API.models.Item;
@@ -10,11 +12,11 @@ import com.projekt.Spring_Boot_API.repositories.IFolderRepository;
 import com.projekt.Spring_Boot_API.repositories.IItemRepository;
 import com.projekt.Spring_Boot_API.repositories.IUserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -24,7 +26,7 @@ public class ItemService {
     private final IFolderRepository folderRepository;
     private final IItemRepository itemRepository;
 
-    public Item uploadItem(String name, MultipartFile file, int size, UUID folderId, UUID userId) {
+    public Item uploadItem(String name, MultipartFile file, int size, UUID folderId) {
         byte[] fileBytes = null;
         try {
             fileBytes = file.getBytes();
@@ -37,13 +39,14 @@ public class ItemService {
             throw new RuntimeException("File has no contents.");
         }
 
-        User user = userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
+        User owner = authenticateUser();
 
-        Folder folder = folderRepository.findByFolderId(folderId)
+        Folder location = folderRepository.findByFolderId(folderId)
                 .orElseThrow(FolderNameEmptyException::new);
 
-        Item item = new Item(name, fileBytes, size, folder, user);
+        checkFolderOwnership(owner, location);
+
+        Item item = new Item(name, fileBytes, size, location, owner);
 
         return itemRepository.save(item);
     }
@@ -52,16 +55,23 @@ public class ItemService {
         Item item = itemRepository.findByItemId(itemId)
                 .orElseThrow(RuntimeException::new);
 
-        if (itemName != null && !itemName.isBlank()) {
-            item.setItemName(itemName);
-        }
+        User user = authenticateUser();
+
+        checkItemOwnership(user, item);
 
         if (itemLocationId != null) {
             Folder folder = folderRepository.findByFolderId(itemLocationId)
                     .orElseThrow(FolderNameEmptyException::new);
 
+            checkFolderOwnership(user, folder);
+
             item.setFolder(folder);
         }
+
+        if (itemName != null && !itemName.isBlank()) {
+            item.setItemName(itemName);
+        }
+
         itemRepository.save(item);
     }
 
@@ -69,18 +79,42 @@ public class ItemService {
         Item item = itemRepository.findByItemId(itemId)
                 .orElseThrow(RuntimeException::new);
 
+        User user = authenticateUser();
+
+        checkItemOwnership(user, item);
+
         itemRepository.delete(item);
     }
 
-    public List<Item> getItemsInFolder(UUID folderId) {
-        Folder folder = folderRepository.findByFolderId(folderId)
-                .orElseThrow(FolderNameEmptyException::new);
+    public Item downloadItem(UUID itemId) {
+        Item item = itemRepository.findByItemId(itemId)
+                .orElseThrow(ItemNotFoundException::new);
 
-        return itemRepository.findByFolder(folder);
+        User user = authenticateUser();
+
+        checkItemOwnership(user, item);
+
+        return item;
     }
 
-    public Item downloadItem(UUID itemId) {
-        return itemRepository.findByItemId(itemId)
-                .orElseThrow(ItemNotFoundException::new);
+    private User authenticateUser() {
+        return (User) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+    }
+
+    private void checkItemOwnership(User user, Item item) {
+        if (!item.getUser().getUserId()
+                .equals(user.getUserId())) {
+            throw new OwnerItemMismatchException();
+        }
+    }
+
+    private void checkFolderOwnership(User user, Folder folder) {
+        if (!folder.getUser().getUserId()
+                .equals(user.getUserId())) {
+            throw new OwnerFolderMismatchException();
+        }
     }
 }
